+++ react_frontend/src/components/adminList.js
