+++ react_frontend/src/components/studentListItem.js
