+++ react_frontend/src/index.js
@@ -1,0 +1,8 @@
+// @flow
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './pages/App';
+
+
+ReactDOM.render(<App />, document.getElementById('app'),);
+
